Refactor visitor repository and service
The visitor repository was updated to use the IDbContextFactory instead of a direct context, improving how database context is managed and consumed. An empty VisitorDto list check was also added in the VisitorService to handle potential situations with no visitors. Relationship mapping in VisitorConfiguration was also adjusted from one-to-one to one-to-many between addresses and visitors, as one address can cater to multiple visitors.