Implement MessageBoxes and PDF creation
The StatisticWindow.xaml.cs file was updated to allow the window to receive various messages, and then display MessageBoxes in response. PDF creation functionality was introduced to the StatisticViewModel.cs file, and error handling was added for this new functionality. Additionally, the Statistic.Wpf.csproj file was updated to use Windows Forms.