Update StatisticWindow.xaml layout and design
Significant changes have been made to the application's StatisticWindow.xaml layout, including adjustments to element alignments and addition of new functionalities. These changes improve the layout of the front-end for a more intuitive user experience. Furthermore, the design changes include adding a color zone, modifying styles for checkboxes, and improving navigation features for usability and accessibility. The overall scope was to clean up the frontend and improve user interaction with the program.