Add handling of database errors on application start
Added handling of errors during fetching of data from the database on application startup in the StatisticViewModel. Now application displays the error message to the user through a MessageBox in the StatisticWindow class. Also removed the logic of ensuring the database from Direct App start and encapsulated it to a service which gets called during window load.